Refactor and fix controller tests
