+++ test/test_models.rb
@@ -135,3 +135,65 @@ def define_routes
     end
   end
 end
+
+class Foo
+  include ActiveModel::Model
+  attr_accessor :id, :is_root_resource, :updated_at, :created_at
+  cattr_accessor :_id
+
+  self._id = 1
+
+  def self.find(*_args)
+    Foo.new.tap do |f|
+      f.id = _id
+      self._id += 1
+      f.updated_at = DateTime.parse('1970-01-01 00:01:00')
+    end
+  end
+
+  def self.inject(*args, &block)
+    self._id = 1
+    [find, find].inject(*args, &block)
+  end
+
+  def self.order(*_args); self; end
+  def self.page(*_args); self; end
+  def self.per(*_args); self; end
+  def self.where(*_args); self; end
+  def self.build; new; end
+end
+
+class FooRepresenter
+  def self.prepare(_model)
+    FooRepresenter.new
+  end
+
+  def from_json(str, opt)
+    Foo.new
+  end
+end
+
+class FoosController < ActionController::Base
+  include HalApi::Controller
+
+  cattr_accessor :_caches_action
+
+  attr_accessor :_respond_with, :request, :params
+
+  def params
+    @params ||= ActionController::Parameters.new(action: 'update', id: 1)
+  end
+
+  def request
+    @request ||= OpenStruct.new('put?' => false, 'post?' => false, 'content_type' => 'application/json')
+  end
+
+  def respond_with(*args)
+    self._respond_with = args
+  end
+
+  def self.caches_action(action, options = {})
+    self._caches_action ||= {}
+    self._caches_action[action] = options
+  end
+end
